AG-17107 Join duplicate SeriesGrouping types
Conflicts:
	packages/ag-charts-community/src/module/optionsModule.ts
