+++ packages/ag-charts-community/src/chart/series/seriesLayerManager.ts
@@ -2,7 +2,7 @@ import { SeriesZIndexMap, clamp } from 'ag-charts-core';
 
 import { Group } from '../../scene/group';
 import { compareZIndex } from '../../scene/zIndex';
-import type { SeriesGrouping } from './seriesStateManager';
+import type { SeriesGrouping } from '../../module/seriesGrouping';
 
 interface SeriesConfig {
     internalId: string;
