+++ packages/ag-charts-community/src/module/seriesGrouping.ts
@@ -1,0 +1,6 @@
+export type SeriesGrouping = {
+    groupIndex: number;
+    groupCount: number;
+    stackIndex: number;
+    stackCount: number;
+};
